git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@976 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/registrartable.c
@@ -796,13 +796,13 @@ int registrarTableGetRegistrar(struct RegistrarTable*   registrarTable,
                                                              assocID);
                   if(sd >= 0) {
                      LOG_ACTION
-                     fprintf(stdlog, "Taking assoc %u as new registrar => socket %d\n", assocID, sd);
+                     fprintf(stdlog, "Taking assoc %u as new registrar => socket %d\n", (unsigned int)assocID, sd);
                      LOG_END
                      return(sd);
                   }
 
                   LOG_WARNING
-                  fprintf(stdlog, "Got crap via assoc %u -> aborting it\n", assocID);
+                  fprintf(stdlog, "Got crap via assoc %u -> aborting it\n", (unsigned int)assocID);
                   LOG_END
                   flags = 0;
                   recvfromplus(registrarHuntFD,
